RBAC changes

+++ src/main/java/com/grievancemanagementservice/controller/GrievanceController.java
@@ -3,20 +3,19 @@ package com.grievancemanagementservice.controller;
 import com.grievancemanagementservice.model.Grievance;
 import com.grievancemanagementservice.service.GrievanceService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.naming.NoPermissionException;
+import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -28,13 +27,17 @@ public class GrievanceController {
     private GrievanceService grievanceService;
 
     @GetMapping
-    public List<Grievance> getAllGrievances() {
-        return grievanceService.getAll();
+    public List<Grievance> getAllGrievances(@RequestHeader LinkedHashMap headers) {
+        System.out.println("Header userId : " + headers.get("userid"));
+        UUID userId = UUID.fromString(headers.get("userid").toString());
+        return grievanceService.getAll(userId);
     }
 
     @GetMapping("/{id}")
-    public Optional<Grievance> getAllGrievances(@PathVariable("id") UUID id) {
-        return grievanceService.getById(id);
+    public Optional<Grievance> getGrievance(@PathVariable("id") UUID id, @RequestHeader LinkedHashMap headers) {
+        System.out.println("Header userId : " + headers.get("userid"));
+        UUID userId = UUID.fromString(headers.get("userid").toString());
+        return grievanceService.getById(id, userId);
     }
 
     @PostMapping
@@ -49,8 +52,14 @@ public class GrievanceController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteGrievance(@PathVariable("id") UUID id) {
-        grievanceService.deleteById(id);
+    public String deleteGrievance(@PathVariable("id") UUID id, @RequestHeader LinkedHashMap headers) throws NoPermissionException {
+        System.out.println("Header userId : " + headers.get("userid"));
+        UUID userId = UUID.fromString(headers.get("userid").toString());
+        if("SUPER_ADMIN".equals(grievanceService.getUserDetails(userId).getUserRole().toString())) {
+            grievanceService.deleteById(id);
+            return "Grievance deleted successfully!";
+        }
+        throw new NoPermissionException("You don't have permission to do this operation");
     }
 
     @GetMapping("/{grievanceId}/{departmentId}")
